mini-python-backend/requirements.txt
antlr-denter==1.3.1
antlr4-python3-runtime==4.9.3
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
six==1.16.0
Werkzeug==2.0.3
